Work on Win64.
Some warnings are silenced in TTString, TTValue, TTMatrixArray, because
they were occuring all the time and making the build log unreadable.
However they should be reviewed carefully (as well as the other warnings
that occur when running the build under Win64).

+++ Foundation/library/includes/TTString.h
@@ -350,7 +350,7 @@ public:
 		substring.reserve(n+16);
 		substring.resize(n);
 		for (i=0; i<n; i++) {
-			substring[i] = (*this)[pos + i];
+			substring[static_cast<int>(i)] = (*this)[static_cast<int>(pos+i)];
 			if (pos+i >= size())
 				break;
 		}
